ubuntu

+++ helper/bastion.sh
@@ -13,6 +13,29 @@ echo "==========================================================================
 
 # curl -sL toast.sh/helper/bastion.sh | bash
 
+OS_NAME="$(uname)"
+OS_FULL="$(uname -a)"
+
+if [ "${OS_NAME}" == "Linux" ]; then
+    if [ $(echo "${OS_FULL}" | grep -c "amzn1") -gt 0 ]; then
+        OS_TYPE="amzn"
+    elif [ $(echo "${OS_FULL}" | grep -c "amzn2") -gt 0 ]; then
+        OS_TYPE="amzn"
+    elif [ $(echo "${OS_FULL}" | grep -c "el6") -gt 0 ]; then
+        OS_TYPE="el6"
+    elif [ $(echo "${OS_FULL}" | grep -c "el7") -gt 0 ]; then
+        OS_TYPE="el7"
+    elif [ $(echo "${OS_FULL}" | grep -c "Ubuntu") -gt 0 ]; then
+        OS_TYPE="Ubuntu"
+    elif [ $(echo "${OS_FULL}" | grep -c "coreos") -gt 0 ]; then
+        OS_TYPE="coreos"
+    fi
+fi
+
+if [ "${OS_TYPE}" == "" ]; then
+    error "Not supported OS. [${OS_FULL}]"
+fi
+
 # localtime
 sudo ln -sf "/usr/share/zoneinfo/Asia/Seoul" "/etc/localtime"
 date
@@ -20,16 +43,27 @@ date
 # update
 echo "================================================================================"
 echo "# update... "
-sudo yum update -y
+
+if [ "${OS_TYPE}" == "Ubuntu" ] || [ "${OS_TYPE}" == "coreos" ]; then
+    sudo apt-get update
+elif [ "${OS_TYPE}" == "amzn" ] || [ "${OS_TYPE}" == "el6" ] || [ "${OS_TYPE}" == "el7" ]; then
+    sudo yum update -y
+fi
 
 # tools
 echo "================================================================================"
 echo "# install tools... "
-sudo yum install -y git vim telnet jq make wget docker
+
+if [ "${OS_TYPE}" == "Ubuntu" ] || [ "${OS_TYPE}" == "coreos" ]; then
+    sudo apt-get install -y git vim telnet jq make wget docker
+elif [ "${OS_TYPE}" == "amzn" ] || [ "${OS_TYPE}" == "el6" ] || [ "${OS_TYPE}" == "el7" ]; then
+    sudo yum install -y git vim telnet jq make wget docker
+fi
 
 # aws-cli
 echo "================================================================================"
 echo "# install aws-cli... "
+
 pip install --upgrade --user awscli
 aws --version
 
@@ -48,7 +82,15 @@ fi
 # kubectl
 echo "================================================================================"
 echo "# install kubectl... "
-cat <<EOF > kubernetes.repo
+
+if [ "${OS_TYPE}" == "Ubuntu" ] || [ "${OS_TYPE}" == "coreos" ]; then
+    sudo apt-get update && sudo apt-get install -y apt-transport-https
+    curl -s https://packages.cloud.google.com/apt/doc/apt-key.gpg | sudo apt-key add -
+    sudo touch /etc/apt/sources.list.d/kubernetes.list
+    echo "deb http://apt.kubernetes.io/ kubernetes-xenial main" | sudo tee -a /etc/apt/sources.list.d/kubernetes.list
+    sudo apt-get update && sudo apt-get install -y kubectl
+elif [ "${OS_TYPE}" == "amzn" ] || [ "${OS_TYPE}" == "el6" ] || [ "${OS_TYPE}" == "el7" ]; then
+    cat <<EOF > kubernetes.repo
 [kubernetes]
 name=Kubernetes
 baseurl=https://packages.cloud.google.com/yum/repos/kubernetes-el7-x86_64
@@ -57,13 +99,16 @@ gpgcheck=1
 repo_gpgcheck=1
 gpgkey=https://packages.cloud.google.com/yum/doc/yum-key.gpg https://packages.cloud.google.com/yum/doc/rpm-package-key.gpg
 EOF
-sudo mv kubernetes.repo /etc/yum.repos.d/kubernetes.repo
-sudo yum install -y kubectl
+    sudo mv kubernetes.repo /etc/yum.repos.d/kubernetes.repo
+    sudo yum install -y kubectl
+fi
+
 kubectl version --client --short
 
 # kops
 echo "================================================================================"
 echo "# install kops... "
+
 export VERSION=$(curl -s https://api.github.com/repos/kubernetes/kops/releases/latest | jq --raw-output '.tag_name')
 wget https://github.com/kubernetes/kops/releases/download/${VERSION}/kops-linux-amd64
 chmod +x kops-linux-amd64 && sudo mv kops-linux-amd64 /usr/local/bin/kops
@@ -72,6 +117,7 @@ kops version
 # helm
 echo "================================================================================"
 echo "# install helm... "
+
 export VERSION=$(curl -s https://api.github.com/repos/kubernetes/helm/releases/latest | jq --raw-output '.tag_name')
 curl -sL https://storage.googleapis.com/kubernetes-helm/helm-${VERSION}-linux-amd64.tar.gz | tar xz
 sudo mv linux-amd64/helm /usr/local/bin/helm && rm -rf linux-amd64
@@ -80,6 +126,7 @@ helm version --client --short
 # jenkins-x
 echo "================================================================================"
 echo "# install jenkins-x... "
+
 export VERSION=$(curl -s https://api.github.com/repos/jenkins-x/jx/releases/latest | jq --raw-output '.tag_name')
 curl -sL https://github.com/jenkins-x/jx/releases/download/${VERSION}/jx-linux-amd64.tar.gz | tar xz
 sudo mv jx /usr/local/bin/jx
@@ -88,6 +135,7 @@ jx --version
 # terraform
 echo "================================================================================"
 echo "# install terraform... "
+
 export VERSION=$(curl -s https://api.github.com/repos/hashicorp/terraform/releases/latest | jq --raw-output '.tag_name' | cut -c 2-)
 wget https://releases.hashicorp.com/terraform/${VERSION}/terraform_${VERSION}_linux_amd64.zip
 unzip terraform_${VERSION}_linux_amd64.zip && rm -rf terraform_${VERSION}_linux_amd64.zip
@@ -97,13 +145,20 @@ terraform version
 # java
 echo "================================================================================"
 echo "# install java... "
-sudo yum remove -y java-1.7.0-openjdk
-sudo yum install -y java-1.8.0-openjdk java-1.8.0-openjdk-devel
+
+if [ "${OS_TYPE}" == "Ubuntu" ] || [ "${OS_TYPE}" == "coreos" ]; then
+    sudo apt-get install -y openjdk-8-jdk
+elif [ "${OS_TYPE}" == "amzn" ] || [ "${OS_TYPE}" == "el6" ] || [ "${OS_TYPE}" == "el7" ]; then
+    sudo yum remove -y java-1.7.0-openjdk
+    sudo yum install -y java-1.8.0-openjdk java-1.8.0-openjdk-devel
+fi
+
 java -version
 
 # maven
 echo "================================================================================"
 echo "# install maven... "
+
 export VERSION=3.5.3
 if [ ! -d /usr/local/apache-maven-${VERSION} ]; then
   curl -sL https://www.apache.org/dist/maven/maven-3/${VERSION}/binaries/apache-maven-${VERSION}-bin.tar.gz | tar xz
@@ -115,14 +170,22 @@ mvn -version
 # nodejs
 echo "================================================================================"
 echo "# install nodejs... "
-curl --silent --location https://rpm.nodesource.com/setup_10.x | sudo bash -
-sudo yum install -y nodejs
+
+if [ "${OS_TYPE}" == "Ubuntu" ] || [ "${OS_TYPE}" == "coreos" ]; then
+    curl -sL https://deb.nodesource.com/setup_10.x | sudo -E bash -
+    sudo apt-get install -y nodejs
+elif [ "${OS_TYPE}" == "amzn" ] || [ "${OS_TYPE}" == "el6" ] || [ "${OS_TYPE}" == "el7" ]; then
+    curl -sL https://rpm.nodesource.com/setup_10.x | sudo bash -
+    sudo yum install -y nodejs
+fi
+
 echo "node $(node -v)"
 echo "npm $(npm -v)"
 
 # heptio
 echo "================================================================================"
 echo "# install heptio... "
+
 wget https://amazon-eks.s3-us-west-2.amazonaws.com/1.10.3/2018-06-05/bin/linux/amd64/heptio-authenticator-aws
 chmod +x heptio-authenticator-aws && sudo mv heptio-authenticator-aws /usr/local/bin/heptio-authenticator-aws
